Initial Commit of files
We are loading and displaying contacts, we applied ThemedText, ThemedView and ThemedTextInput to all display elements.
We are utilizing FlatList to display large amount of contacts for now, there is no continuous scroll ability at the moment, once we hit the end of the partial list there is a pause until the rest of the list loads.

TODO:
Center the contact name display in the ListItem
Add the WhiteListing Feature so that contacts whitelisted will be able to ring through even on Do not Disturb mode.
Add Development Roadmap